Break out `Device` class into it's own file
